Infer: add `from __future__ import annotations` by default
Summary:
One of the most common infer bugs is that we will often annotate methods of a
class as returning that class, without quoting the annotation.

To solve the problem, which can occur in other ways that are harder to detect,
in a general way either requires a fairly powerful algorithm to determine
whethere we're trying to use an incomplete type, or one of two brute-force
solutions:
- quote all annotations - I'd like to avoid this because it's hard to read
- use `from __future__ import annotations`

I decided it makes more sense to go with the latter, and put a flag in to
disable it. The flag is useful in python 3.6 where no such behavior exists
(although in that case infer will continue to be buggy) and in python 3.10+
where no-evaluate-annotations is the default.

Since most companies are not yet using 3.10 and the `from __future__ import
annotations` does no harm for newer python versions, we add the import by
default.

Reviewed By: shannonzhu

Differential Revision: D29850797

fbshipit-source-id: 8359df4aac1a3613d8f59bcd849293712452f6c4