Infer bugfix: recursively reject undefined
Summary:
We were using `Type.is_untyped`, which checks for Top, Bottom, and Any
at the the top-level, and then we were recursively checking for Any and
Top but not Bottom.

This is a problem because (somewhat surprisingly to me) pyre infers
Bottom rather than Top for any parameters that come in without
annotations and get returned without any intermediate operations.

As a result, in cases where the Bottom happened inside of another
type (the most obvious example is a tuple), we were failing to exclude
it and then producing literal "undefined" annotations in codemeods.

This diff removes the untyped check, which is not needed if we
explicitly do recursive checks for Any/Bottom/Top, and adds a recursive
check for Bottom.

Reviewed By: shannonzhu

Differential Revision: D29827579

fbshipit-source-id: 15b979209d88c9dda7a876b365cd815e8de2a468