Infer: verify that working directory is the local or project root.
Summary:
At the moment infer has some brittle path handling because of the need to
translate amongst
- relative paths passed in by hand, if modules to annotate is given
- paths in the typing directory
- paths to actual python files

infer is (somewhat) well-tested when run either from the local root or the project
root, but other cases are likely to produce errors. So instead of running the
expensive analysis and failing with a runtime error afterward, we should check
that the the working directory is supported before doing anything and give
back as useful an error message as possible.

Reviewed By: shannonzhu

Differential Revision: D29853825

fbshipit-source-id: 235cb68526563a1ad23766aa21873f95230fdaf6