Prevent stack overflow in edge case
Summary:
We recently discovered an edge case that can cause pysa to
stack overflow in callgraph generation.

The problem happens as follows:
- when resolving a function call, we check whether the arguments
  might be callable to ensure that we'll detect taint flows due
  to higher-order functions
- now consider what happens if we resolve attribute access for
  any class `A` which has a `__getattr__` annotation which is a union
  type including `A`:
  - we'll check to see if the attribute is callable
  - one of the cases we'll have to handle is when `attribute` is
    itself of type `A`, at which point we want to know if `A` is
    callable
  - the check for a `__call__` method falls back to `__getattr__`
  - now we're back to checking `A` for whether it's callable
  - ... this creates an infinite loop

I'm reasonably sure that the existince of this infinite loop was
known, the `resolve_callees_from_type` prevents us from hitting it
in the case where `__getattr__` returns an `A` directly. But we
were failing to propagate it to resolution of the `__getattr__`
return type in the case where that type was a parametric type
including `A`, as opposed to just a plain `A`.

I also added a check to make sure that if __getattr__ can return
some other class, this logic wouldn't cause us to fail to detect
the possibility that `__call__` actually resolves to some other class's
`__call__` implementation via `__getattr__` magic, which would be
trippy but is possible in python.

Reviewed By: dkgi

Differential Revision: D29772630

fbshipit-source-id: aad705e0cbcc1612a23d12a2eef23383ea32ad34