Add unit tests for TypeAnnotation
Summary:
We have a little bit of logic in TypeAnnotation now (dequalify, prefix)
and we need to add some more as part of a quickfix for an infer bug.

So it makes sense to set up a unit test harness. This commit has
no prod code change, just verifying existing behavior.

Reviewed By: dkgi

Differential Revision: D29900729

fbshipit-source-id: a8682ba5b699bcef2e55473c83aec5ab7515ac2a