Infer refactor: extract StubGenerationOptions
Summary:
I already noticed a pretty ugly antipattern in stub generation
where, because we need the `dequalify` flag at the very lowest
level of the tree when printing an annotation, we have to pass
it down through every single to_stubs method.

This is pretty nasty thing to do, and I realized that to push
out some quick fixes of common infer bugs we're going to want
more flags.

So, this commit preemptively extracts flags (at the moment
they are `use_future_annotations` and `dequalify`) into a
new dataclass `StubGenerationOption`, which allows us to ensure
that anything needing it can get it easily in the construction
of `ModuleAnnotations`.

There's no behaviour change here.

Reviewed By: grievejia

Differential Revision: D29900628

fbshipit-source-id: 095169dcac10a83b6959dc4932c28ec8c966c42f