Fix infer commanline arg nits
Summary:
While working on a bigger refactor, I realized that the order of arguments
was mismatched between the click args, the `infer` function args, and the
`Infer` constructor call. This makes it a little harder to keep track of
things, so fixing it.

Also fixed a typo in one of the click help functions

Reviewed By: grievejia

Differential Revision: D29901067

fbshipit-source-id: 52dac7f1235f1995e96d77296962a1f6cc6f7e08