Infer bugfix: don't produce junk imports for given annotations [1]
Summary:
## Problem

Pyre infer is a long way from being able to handle imports in a
reliable way - it's a difficult problem to reliably deal with
relative and aliased imports, especially aliased ones, and as a result
we frequently produce duplicate or even incorrect imports that break
the linter and make our codemods useless.

This problem is hard to fix outright, but it's especially bad because
we generate stubs that include all the *existing* annotations (in particular
parameter annotations), which means we'll often produce codemods that
have junk imports which weren't even needed to add an annotation.

## Quickfix

The quickest fix is to switch from treating existing annotations as actual
types to using just expressions.

This is a hack, not a nice solution:
- It's gross for a few reasons...
  - it's the wrong thing to do conceptually (when I tried to do this a while
    back, reviewers did not like using expressions because it's in some
    sense less correct, or would be if the rest of our code were reliable)
  - it means the generated stubs are incorrect and in fact misleading, because
    we add literal expressions without any of the import context needed
    to make sense of them
- But it actually does work: the codegen step where we merge stubs and code
  applies whatever annotations are on the stub blindly. As a result, having
  "wrong" annotations with no import in the stub file is just fine in terms
  of the end effect - when the annotation gets added to code the import
  is there and we get a no-op, which is exactly what should happen when
  we're dealing with a given annotation.

## This commit

Fix the problem, prove that we get the desired behavior in our test
of an aliased import.

Reviewed By: dkgi

Differential Revision: D29913447

fbshipit-source-id: 9bc9f2efc5da5a9e931f9381fa724a4a79fb2aff