Infer: refactor error -> result logic [2] validate and sanitize
Summary:
This is the final pure-refactor diff rearranging code to more
cleanly handle three seprate concerns for inference updates
- logic to convert analysis errors to actual inference updates
- plumbing to actually update the data structure
- logic (where we now there are currently bugs to fix) to
  validate and sanitize the type itself

The previous two commits had teased these three concerns apart,
but there was a remaining issue, that `type_to_string` was
doing some sanitization (convert escaped free variables to any
and do the infer transform) at print time, which was quite
confusing.

Why was this confusing? Because I've been trying to debug
cases where `Variables` leak into our infer output, and for
a long time I thought the problem might be that
- we convert free variables to any and then check for any
  in the `ignore` function
- but then we go ahead and save the original type with free
  varaibles!

It turns out this is not the case but that's much, much
easier to see if we keep the sanitize and validate steps
in the same place.

Reviewed By: dkgi

Differential Revision: D29899233

fbshipit-source-id: 47fcf396e163ec1fc94c9bcf1d0bcc05d62c0ab2