Infer bugfix: don't hand back `Variable` annotations.
Summary:
The root cause of the bug was that when sanitizing types, we were converting
all *escaped* free variables to Any but we were not first marking all free
variables as escaped. As a result, we were regularly leaking free variables
into infer output, which is never correct given the current code-generating
logic, there's no case were we could handle a variable correctly.

(Someday maybe we could extend infer to handle generics, although truth be told
that's likely to be low-payoff since generics only exist in codebases that are
already typed).

Reviewed By: dkgi

Differential Revision: D29899234

fbshipit-source-id: e5d40a7de7542e88a72ef7148692eb8479fed0fd