Infer: refactor error -> result logic [0] add Inference
Summary:
This is part of a two-commit stack where I try to separate out 3
bits of logic that are intermingled in the
add_missing_annotation_error function of LocalResult:
- figuring out what kind of update should happen, if any, based
  on an error and the context. This is mostly a very simple pattern
  match.
- actually updating the data. This is the bulk of the code, but is
  again very simple and unlikely to contain bugs.
- running the business logic to validate and sanitize a type, which
  is tricky and I've discovered multiple bugs in this in the last
  week.

This first commit separates the data update from the type mangling
and error matching logic; I'll separate the latter two in the next
diff.

The strategy I'm taking is to have a very simple variant type
used to communicate between the inference logic and the LocalResult
data. I'm calling it `Inference.t` since it represents a single
inference about a single type but I'm open to other suggestions;
I tried to come up with a name involving `Update` but they all
sounded pretty awkward to me. (`InferenceUpdate`? Feels a bit icky)

**Motivation**

I have two motivations here:
- short term: I want to fix a bug where type variables are leaking
  into the inferred annotations, which is harder to do without the
  type mangling cleanly separated.
  - Also, I'm almost certain that we need to replace our existing validation
    (`ignore`) with a sanitize-and-validate step that can rewrite the type as
    well as validate it, which is awkward in current factoring of the code but
    would be easy if we could just map a sanitize-and-validate that returns an
    optional type over the updates
- medium term: we don't want to be converting AnalysisError.t into
  inference results in the long run, we want to directly update the
  inference data instead. The first step of this refactor will be moving
  the error-to-inference logic out of `TypeInferenceData` and into
  `TypeInferenceResult`, which isn't possible if that isn't decoupled
  from the data updates.

Reviewed By: dkgi

Differential Revision: D29899235

fbshipit-source-id: 9ab7f2d8d55a683237049bded01899cd63895841