Infer: refactor error -> result logic [1] validate Inference
Summary:
This is part of a two-commit stack where I try to separate out 3
bits of logic that are intermingled in the
add_missing_annotation_error function of LocalResult:
- figuring out what kind of update should happen, if any, based
  on an error and the context. This is mostly a very simple pattern
  match.
- actually updating the data. This is the bulk of the code, but is
  again very simple and unlikely to contain bugs.
- running the business logic to validate and sanitize a type, which
  is tricky and I've discovered multiple bugs in this in the last
  week.

The previous commit separated the type logic and error handling
from the result updating. This diff extracts the type logic from
the error handling by moving validation into `Inference`, where we
can isolate all the type logic, which is where most infer bugs live,
and keep it separate from the context-and-error logic which needs
to eventually get pushed down into the fixpoint.

In order to enable a future where we might not just reject certain
types but also transform them, I made this an option-producing function
so that we can bind the raw `Inference.t` to the validate function;
it will give us `None` whenever the inference isn't usable and a
(potentially transformed) inference whenever it is.

Reviewed By: dkgi

Differential Revision: D29899236

fbshipit-source-id: d1595de4d1194342e648b335837d6dfcc8af9a35