TypeInferenceData cleanup [1]: Enforce validation of inferred types
Summary:
**This Stack**

I'm working on moving all logic not directly related to building
up type inference information out of `typeInferenceData` and
into `typeInferenceLocal`.

This is what we need to do if we want to
- eventually decouple the internals of type inference from
  `AnalysisError.t`
- handle context information inside the fixpoint, where it
  belongs (for historical reasons some fo that wound up in
  the wrong place)

**This Commit**

This commit has no meaningful logic change, but it prepares
for moving the `AnalysisError` to `Inference` logic out of
`TypeInferenceData` and into `TypeInferenceLocal` by ensuring
that no one can attempt to pass an un-validated inferred type
to `LocalResult`.

We do this as follows:
- define `Inference.raw` to be a `Type.t * target` (which is
  visible in the interface)
- redefine `Inference.t` to be a `raw option`, and do *not*
  make that visible outside of `TypeInferenceData`. The only
  way to construct this is using the validation function, which
  we've renamed to `create`
- tweak `add_inference` to take an option rather than just
  a `Type.t * target`, and in the interface declare it to accept
  the obscure type `Inference.t`, whcih can only be created via
  `create`.

This guarantees that no caller outside of `TypeInferenceData`
could accidentally write code that adds an unvalidated type
to `LocalResult.t`.

Reviewed By: dkgi

Differential Revision: D29957614

fbshipit-source-id: 42746a89e5f0f804280a0af16f323b138346a379