TypeInferenceData cleanup [0]: Add typeInferenceLocal.mli, hide type details
Summary:
**This Stack**

I'm working on moving all logic not directly related to building
up type inference information out of `typeInferenceData` and
into `typeInferenceLocal`.

This is what we need to do if we want to
- eventually decouple the internals of type inference from
  `AnalysisError.t`
- handle context information inside the fixpoint, where it
  belongs (for historical reasons some fo that wound up in
  the wrong place)

**This Commit**

Add an interface file to `TypeInferenceData`, which will
allow us to do the rest of the refactor using implementation hiding
techniques to get better type safety.

I deferred addding an interface file for a long time because the
infer data handling wasn't very stable, and I found that juggling
interface files of immature code was difficult, particularly when
addressing code review comments in stacks.

But at this point the infer data structure is pretty stable and it's
time to have an interface.

I used my trick of first trying
```
opam install ocaml-print-intf
dune exec -- ocaml-print-intf interprocedural_analysis/type_inference/typeInferenceData.ml
```
to bootstrap the interface, but this spits out a *lot* of junk
and I deleted most of it because I'm actually obscuring almost all
the data.

In order to support this change, I made to additional tweaks:
- in `typeInferenceLocal.mli`, use `Ast.Reference.t` rather than
  `SerializableReference.t`; they are the same but we're no longer
  exposing that.
- add a `define_name` method to `LocalResult.t` so that tests can get
  at the name without pattern-matching; this is part of what lets
  us keep the data hidden

Reviewed By: dkgi

Differential Revision: D29955184

fbshipit-source-id: 0aa289dc83d8b6bab8fb9adc2632d8082b651b9b