TypeInferenceData cleanup [2]: Move AnalysisError -> Inference logic
Summary:
**This Stack**

I'm working on moving all logic not directly related to building
up type inference information out of `typeInferenceData` and
into `typeInferenceLocal`.

This is what we need to do if we want to
- eventually decouple the internals of type inference from
  `AnalysisError.t`
- handle context information inside the fixpoint, where it
  belongs (for historical reasons some fo that wound up in
  the wrong place)

**This Commit**

This is the core commit of the stack, moving the AnalysisError.t
to Inference.t conversion logic to the module where it needs to live
in order for us to eventually kill it.

There's a bit of messiness in the way we had to add an `abstract` function
to the interface; the abstract check also needs to also move because it's
part of the context rather than the actual inference result I will clean
this up in the next commit.

Reviewed By: dkgi

Differential Revision: D29957616

fbshipit-source-id: cc1f064e04f9e10849c4a2de79958f4dfe89fda5