TypeInferenceData cleanup [3]: Get rid of `abstract` field in `LocalResult.t`
Summary:
**This Stack**

I'm working on moving all logic not directly related to building
up type inference information out of `typeInferenceData` and
into `typeInferenceLocal`.

This is what we need to do if we want to
- eventually decouple the internals of type inference from
  `AnalysisError.t`
- handle context information inside the fixpoint, where it
  belongs (for historical reasons some fo that wound up in
  the wrong place)

**This Commit**

The `abstract` field was always an ugly thing to have in our inference result
type, but it was necessary at first because of the way the old
`AnalysisError.inference_information` logic was structured.

But in reality `abstract` is just a part of the inference context (it depends
solely on the define signature), and as such it belongs in
`typeInferenceLocal.ml`, where we can eventually push it into the fixpoint
context.

That's what this commit does. This is the final commit of a stack where I clean
unwanted logic out of `typeInferenceData.ml` by movng it into
`typeInferenceLocal.ml`; at this point we are ready to start refactoring
`typeInferenceLocal.ml` but there are no more module migrations needed
to have the infer backend in a reasonable state.

Reviewed By: dkgi

Differential Revision: D29957615

fbshipit-source-id: fdbbb841e6395f5a0e2427ee6bf9fdcfa61ab388