Clean up copypasta in Infer tests
Summary:
It's been bothering me for a while that I have to list out all of
the many infer flags for every unit test that uses an Infer
command.

I like that we don't set default values in the in the constructor, to ensure
that the click wrapper is using all arguments, but it makes sense to have
a helper method in the tests with default values.

Reviewed By: pradeep90

Differential Revision: D29938801

fbshipit-source-id: c9a5b666223653458a06b2e5069263831d3af398