Changed `local_replace` to support substitution for variadic tuples in `Compose`
Summary:
+ Problem: We cannot substitute for unpacked variadic tuples within a `Compose`.
+ Background: This is one of a series of diffs on implementing `Compose`, a type operator intended to take in things like `Callable`s and synthesize the type of their function composition. See D29803910 for details. We previously implemented some degree of normalization for literal annotations, but we do not yet do any substitution inside of a `Compose` - this is a critical feature necessary for `nn.Sequential`, since it has type `Compose[*Ts]`. So we need to be able to do this, so we can type `Sequential` models and stuff.
+ Solution: This diff adds in some clauses to `local_replace` to do normalization. For starters, if you see a variadic, then you can just substitute no problem. The bigger issue comes with what you do after. As of the previous diff, there's a "normal form" for `Compose` - it is a `Concrete` of `Concatenation`s, potentially - this means that anything like `Compose[*Ts]` is really `Compose[Compose[*Ts]]`.
This, however, means that if you do the obvious simplification (`Compose[*Ts] => Compose[some, stuff, here]`), then in the prior case it still won't be at the "top" level (`Compose[Compose[*Ts]] => Compose[Compose[some, stuff, here]]` instead of `Compose[Compose[*Ts]] => Compose[some, stuff, here]`).
**Why not just make the previous diff simplify `Compose[*Ts]` to `Compose[*Ts]`?**
Yeah, you could do that. Then `Compose[int, *Ts, str]` is just that, and not actually `Compose[int, Compose[*Ts], str]`. The thing is, this only works in the case where there is one unpacking - in the more general case, you have no choice but to "quarantine" each unpacked thing on its own, because a `Concatenation` can only have one.
The change made in this diff is, if you have a `Compose` in any capacity, simplify it (and thus the possible redexes from prior `local_replace` calls). This can cause a degree of redundancy, but it's your only choice when multiple variadics are involved, because otherwise you don't know where the variadics are, and you need to bring their elements to the top level.

I think it's better to standardize the process and omit the special casing, since you have to do this for the general case anyways. It's also much simpler, and allows you certain invariants (like the fact that a `Concatenation` never shows up with a prefix and suffix).

Open to discussion.

Reviewed By: pradeep90

Differential Revision: D29911806

fbshipit-source-id: 17dd3a16186eecdf171fefc299d7cfd11260bbb9