Added basic support for `Compose` composition mechanism
Summary:
+ Problem: Currently, anything of type `Compose[...]` is never actually "solved" to find a final composed type, which means that it always causes a type error on being invoked. We don't want that.
+ Background: See D29803910. We would like for `Compose` to, when invoked on an argument, synthesize the "composed" type of its type parameters, and then try to apply it to the argument. We first need to implement this "composition" mechanism, however. This will let us implement `nn.Sequential`, which exhibits this composition behavior.
+ Solution: We added in `compose_list`, which does the actual composition. It lives in a new file, to make sure that `type.ml` is non-contextual. Also, because of a thing where `signature_select` returns free but escaped variables that were given as arguments, it's necessary to bind the variable given as input at the start, and unbind it at the end.

**Why not have this behavior occur eagerly (in `local_replace`)?**
So basically, why not delay it until `forward_callable`. The problem is that signature selection (finding the type of a second function applied to the result of a first function) is contextual, and relies on `GlobalResolution`. `local_replace` is in `type.ml`, which has no such global resolution, so we can't do it there. So we put it off until we absolutely have to, when the object is called.

This is a weak solution for now - we end up binding and unbinding all variables, which will present an issue when there are other existing bound variables. Work to come.

Reviewed By: pradeep90

Differential Revision: D29858921

fbshipit-source-id: af9909d3c54650f40641776e8f4ffff36c01351d