Added `TypeOperation` variant to `Type.t` for `Compose` type operator
Summary:
+ Problem: We would like to support a type for `nn.Sequential`. The best way to do that is with a new variant to `Type.t` for type operators. The reasoning for that is below.
+ Background: We are trying to make changes to Pyre that are more amenable to supporting tensor shape typing. A widely used layer in PyTorch is `nn.Sequential`, which takes in multiple layers and returns another layer, which has the action of their composed transformations. This change will let us annotate more PyTorch code, which will further demonstrate the usefulness of tensor shape typing.
+ Solution: This diff adds in a `TypeOperation` variant to `Type.t`, which is intended to be populated further by any operations which can be done on types, that may return a result type, or may be unable to be performed due to lacking information.

**What should the type of a `Sequential` be?**
To ascertain this, we have to look at a use case. Consider:
```
layer = nn.Sequential(
    nn.Sequential(
        nn.Linear(10, 20),
        nn.Linear(21, 30)
    )
    nn.Linear(30, 40)
)
```
In this case, `Sequential` `__init__` method gets the types of the first `Sequential` and the third `Linear` as arguments. The type of a `Linear` is already determined to be `Linear[L[30], L[40]]` - in other words, a `Parametric`. Note that `Sequential` can only take in objects which subclass from `nn.Module`, or else it results in a run-time error. This means that we cannot have `Sequential` be a `Callable`, or else we would have to have `Sequential` able to take in arbitrary `Callable`s.

A given `Sequential` needs to encode the information about the layers that it contains, and the operations they need to perform. The key thing is that a `Sequential` need not already perform the composition of its layers, that only happens when it's called. The above program does not result in a run-time error, because `layer` is not used. So we should push off "composition" until a later point, when the layer object is used.

How do we retain the information about the layers in the operator long enough until we get to composition? We can do that by just using the facility already available to us - the types of the arguments. From the types of the arguments, we can obtain the `__call__` methods on the object (this is already done elsewhere in the code). So we will just have a `Sequential` object of the same type as the class - a `Parametric` named `Sequential`.

**Why add a new variant to `Type.t`?**
We need a type to represent a `Compose` which has not yet finished. Consider the `Compose[*Modules]` example above. Since we do not have enough information to say what this should resolve to (due to the free type variable), we need to give it some other type (which can be substituted into at a later point), but is still essentially a "frozen" `Compose`. So we need some type representation in `Type.t`.

Adding a `TypeOperation` clears things up in several ways. For one, a type operation (in general) has some common properties - for instance, it should not be able to be matched with other types. Adding a variant for `TypeOperation`s in general means that they can be handled separately, rather than separating their logic.

We can't use `Callable`, because its type representation currently does not allow for any "freezing". We have to keep the layers of the `Compose` separate until it is time to combine. For our purposes, it's better that we use a new variant as opposed to a `Parametric`, to avoid any more hard-coding based on the name of the `Parametric`.

**Why introduce a type operator?**
Given that a `Sequential` object is of type `Sequential[Linear[L[10], L[20]], Linear[L[20], L[30]]]`, for instance, then we need to recover the correct type for its `__call__` method. This can be done by directly having the following stubs:
```
class Sequential(Generic[Unpack[Ts]]):
    def __init__(self, layers: Unpack[Ts]): ...
    __call__: Compose[*Ts]
```

This means that any `Sequential`-like class can annotate the correct logic without needing to be special-cased by Pyre. It also means that we can use the information available in the `Ts` (the layers of the `Sequential`). `Compose` will be implemented in Pyre such that it will try to compose the `__call__` methods of the types it is given, and return either a composed type (a `Callable`), an error, or it doesnt' have enough information to compose (free variables), meaning it stays a `Compose`.

We need the type operator because we have to have a way of signaling to Pyre to do the composition on some variadic tuple of things. Currently, there is no possible annotation we could give to `nn.Sequential` in order to make it do the correct thing, so we need `Compose`.

**Why not add a variant to `Callable`?**

`Callable`'s current representation is pretty verbatim just that of a resolved `Callable`, which means we'd have to add a completely separate variant that looks very different. This would cause us to have to refactor code approximately three miles downstream, which is far more trouble than it's worth for this one feature. It's best to keep `Callable` the way that it is.

**What are the inputs and outputs of the operator?**
`Compose` is parameterized over `Ts`, a variadic tuple. While we don't have support for bounds for variadic tuples at the moment, this is a planned feature, and should be used when it is available. For the time being, `Compose` is intended to take in either `Callable`s, or objects which have associated `__call__` methods. It outputs an object of type `Compose[Ts]`, which, when called on an argument, runs the composition process on its arguments and determines if it type-checks. In the future, once we have bounds on variadic tuples, the definer of the `Sequential` model should put a bound on the `Ts` that they parameterize `Compose` by so that it can only be given objects that subclass `nn.Module`, which is the current behavior of `nn.Sequential`.

Reviewed By: pradeep90

Differential Revision: D29803910

fbshipit-source-id: 54a351a17ca4c927cd9ad6327e7ab4b095f46011