Added support for nested `Compose` operators
Summary:
+ Problem: We currently do not simplify nested `Compose`s in any way whatsoever.
+ Background: Currently, we do not do anything about nested `Compose`s. This means that `Compose` operators can potentially be handed off to `get_call_attribute`, which naturally has no idea what you're talking about. This is bad.
For context on what's going on, see D29803910. We're trying to add a `Compose` type operator, which can take in `Callable`s or things that fulfill the `Callable` protocol, and finds the function type of their composed result. This is helpful for `nn.Sequential`, which we cannot otherwise type.
+ Solution: This diff adds in nested support, by "expanding" all of the elements of a given record, and concatenating them using the `OrderedTypes.concatenate` function already provided. This means that we enforce that a `Compose` should only contain one unpacked variadic.

TODO: T96555096
We may loop forever because of recursive typing of `__call__` methods

Reviewed By: pradeep90

Differential Revision: D29911805

fbshipit-source-id: 1ca75bc3bb65d1a94c72be2d2dc7a30906834e42