Infer bugfix: ModuleAnnotations.is_empty should respect annotate_attributes
Summary:
In D29822289 (https://github.com/facebook/pyre-check/commit/9bbdedada69a04ca0a86ab1eca8ad2771d306732), I added a flag to control whether we attempt to
annotate attributes, which for the most part we cannot do
correctly enough to be useful, with a default of False.

But I made an error in the commit: I pushed the handling of
that flag to the stub *generation* stage of `ModuleAnnotations`,
rather than the construction.

The problem with this is that it causes `ModuleAnnotations.is_empty`
to return false even when there are no annotations other than
attributes, and this can result in junk codemod diffs that do nothing
but modify the imports.

Reviewed By: grievejia

Differential Revision: D29976999

fbshipit-source-id: 71db3170f4dc1655b24fe0cece1a4a4f587fe14f