Patch flaky infer unit tests
Summary:
One of the infer unit tests failed to set `print_only`, which meant
it would trigger filesystem operations. This was generally okay in
the sense that the tests pass, but it makes tests slower and
potentially flaky.

This commit sets `print_only=True` in the `make_command` method so
that it's no longer possible to accidentally fail to set it in a
unit test.

Reviewed By: grievejia

Differential Revision: D29991490

fbshipit-source-id: 50c4a372982939036287b6802eba89536830bf87