Ensure globals with literal values are considered annotated
Summary:
We don't require globals with literal values or otherwise simple inferrable values to be explicitly annotated.

We therefore should also not count these as unannotated when gathering coverage statistics.

RFC: this diff only addresses literal number and string values. There are simple expressions that we are also able to resolve, but libCST is not able to differentiate simple names. We could either err on the side of over- or under-counting these.

Reviewed By: dkgi

Differential Revision: D30061429

fbshipit-source-id: 4a370230ab89eccc8ab56138759ace287b71f99b