Allow setting arbitrary attributes on classes with __setattr__ and __getattr__
Summary:
This diff tries to tackle the issue of when to allow setting arbitrary attributes on a class. In this case we define "arbitrary attribute" as attribute names that aren't declared in the constructor or toplevel.

**tl;dr**: we should only allow setting arbitrary attributes on a class that has both a `__setattr__` defined that accepts the right type (which applies to most classes, since `object` defines a `__setattr__` accepting `Any`), and has a `__getattr__` set so arbitrary attribute access is also not a problem and enforcing the order of setting/accessing of attributes is not a potential source of unsoundness.

**Longer summary**:

The existing behavior was quite inconsistent and confusing, where only classes with an explicit `__setattr__` method with a `value: Any` parameter would permit setting of arbitrary attributes.

Any subclasses that inherited a `__setattr__` from a parent would not allow this, ie.:

```
from typing import *

class Foo:
    def __setattr__(self, name: str, value: Any) -> None: ...

class Bar(Foo):
    pass

class Baz(object):
    pass

def test(foo: Foo, bar: Bar, baz: Baz) -> None:
    foo.attr = 1 # no error
    bar.attr = 1 # error
    baz.attr = 1 # error
```

This is confusing for two reasons:
1. `__setattr__` is commonly inherited, so the discrepancy between `foo` and `bar` above is unintuitive
2. All classes actually have `__setattr__` defined, it's even provided in the typeshed builtins.pyi `object` class. This is because at runtime, you are allowed to set arbitrary attributes on classes. If they don't exist yet, they'll be initialized and then set.

The existing implementation also had unintuitive behavior on the `__getattr__` side:

```
from typing import *

class Foo:
    def __getattr__(self, x: Any) -> Any:
        return x

def test(f: Foo) -> None:
    access = f.some_attribute  # no error
    f.some_attribute = 1 # Attribute `some_attribute` is not defined! Even though it's the same attribute as the line above!
```

There's really only one reason we might want to error, ever, on setting arbitrary attributes on classes. (Of course, besides the edge case where someone overrides the default `__setattr__` and does something unexpected with it, but we aren't responsible for weirdness if you explicitly break assumptions about what this special method does).

That reason is to avoid surprising the user when they go access an attribute and find it isn't defined. This is the same reason we bother requiring users to declare and initialize attributes in the toplevel or the constructor. Otherwise, we can't guarantee that for all instances of this class, the method that instantiates the attribute will be called before a method that reads from it, and we will not be able to know whether reads of any attributes are safe.

So by the same logic, setting arbitrary attributes on classes should not be restricted on the precense of `__setattr__`, but on the presence of `__getattr__`. The latter guarantees that an undeclared attribute that's arbitrarily set in some method can also be arbitrarily accessed on this class, regardless of interprocedural call order, etc.

This diff (1) loosens the `__setattr__` requirement so it can be inherited from a parent, essentially making all classes meet this requirement, (2) stops checking for a precise `Any` type on the `value` parameter of `__setattr__` - any super type of the value being set is fine, (3) requires `__getattr__` to be defined or inherited on any class where we permit arbitrary attributes to be set with no warning.

Reviewed By: grievejia

Differential Revision: D30090379

fbshipit-source-id: 7d344d829a79f699638d4a93361f383228f859b0