Adds taint sources for wtforms and flaskwtf
Adds taint sources for wtforms and flaskwtf. The way either of the two
libraries are used is by inheriting the Form class by a user defined
class. Since, pyre can't find the data attribute of the Form or the
Field (they are created dynamically), tainting all attributes of classes
inheriting from Form class of flaskwtf and wtforms.

Signed-off-by: Abishek V Ashok <abishekvashok@gmail.com>