Consider simple names and calls as annotated in global statistics
Summary:
We are showing warnings on user diffs due to increases in unannotaed values when users add globals that can be implicitly inferred by the type checker.

For example:

```
x = 1
x = "literal string"
x = MyClass()
x = AnotherGlobal
```

These are all designated as implicitly annotated in the type checker, and strict mode will not complain that things like this are not annotated.

However, our statistics script that gathers annotation counts through the codebase is libCST based, and we look for AnnotatedAssign vs. Assign nodes.

The diff below this fixed the easy case, where literal numbers and strings in the value slot of an unannotated assign libCST node can be counted as annotated for Pyre.

This diff is making a harder tradeoff -- we can only distinguish nodes of type basic name or call, and some of these (ie., the 3rd and 4th example lines above) are implicitly annotated, but some may indeed be globals in need of annotation, like a call to an arbitrary function.

See libCST Nodes documentation: https://libcst.readthedocs.io/en/latest/nodes.html#names-and-object-attributes

RFC: I think the tradeoff is probably worthwhile here, ie., over-approximating globals of this style that are implicitly annotated to avoid creating confusion where users are trying to identify why they're getting reduced coverage warnings on their diffs and don't see what they still need to annotate. We can still rely on strict mode to throw errors on globals that can't be interpreted at env building time.

Reviewed By: dkgi

Differential Revision: D30073669

fbshipit-source-id: 0f08467f7a3d983656e2d1eca7bfdd5bc234f277