Add handling for pushing state to children in transform
Summary:
Our current transform visitor handles all of the recursive calls on expressions within expressions within statements within statements, etc.

I'd like to be able to reuse this to implement a preprocessing step that mangles private attributes inside classes, but discovered we lack basic support for passing information along from a parent into children when scheduling recursion into child nodes.

We do keep a state around, which is assigned to by reference and is a way to keep custom aggregate information around throughout the transform.

This diff simply permits the function that decides whether or not to recurse into the children of a node to also make alterations to the state if it wants to. That state is then set before the recursion is scheduled.

The option to say, revert the state or make changes after all the recursed children have been popped off the stack is already present, as the transform of any statement itself is given a chance to modify the state after all of its children are finished.

Reviewed By: dkgi

Differential Revision: D30123395

fbshipit-source-id: 91143b5f6985a0912a25b197db6f5c037392bed2